requirements.txt
python-telegram-bot==22.6
Pillow==11.3.0